requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
future==0.18.2
numpy==1.19.2
pycparser==2.20
python-telegram-bot==12.8
six==1.15.0
torch==1.6.0+cpu
torchaudio==0.6.0
tornado==6.0.4
